fix: number cast mode

Created-by: huang-jz
Commit-by: huang-jz
Merged-by: turing_project1
Description: ## 描述
<!--在这里详细描述你的改动，包括改动的原因和所采取的方法。-->
通过将数值转换为uint64类型，解决数值截断风险。

## checklist
<!--  [x] 表示选中 -->
- [x] 是否通过本地IDE对代码进行静态检查
- [x] 是否通过本地IDE对代码进行格式化处理
- [x] 是否进行空指针校验
- [x] 是否进行返回值校验
- [x] 是否正确释放new/malloc申请的内存
- [x] 是否充分考虑接口的异常场景
- [x] 是否正确记录错误日志

See merge request: cann/ascend-transformer-boost!1598

+++ src/kernels/mixkernels/swi_glu_quant/op_kernel/swi_glu_quant.h
@@ -156,7 +156,7 @@ private:
 
         LocalTensor<float> scaleLocal = scaleQueue.DeQue<float>();
         DataCopyParams copyParams1{1, (uint16_t)(basicRowLenCal * sizeof(float)), 0, 0};
-        uint32_t oooo = baseRow + basicRowLen * ridx;
+        uint64_t oooo = static_cast<uint64_t>(baseRow) + static_cast<uint64_t>(basicRowLen) * static_cast<uint64_t>(ridx);
         DataCopyPad(scale_Gm[oooo], scaleLocal, copyParams1);
         scaleQueue.FreeTensor(scaleLocal);
     }
